+++ src/Renderers/DefaultRenderer.php
@@ -19,7 +19,7 @@ use MarcinOrlowski\AsciiTable\AsciiTable;
 use MarcinOrlowski\AsciiTable\Cell;
 use MarcinOrlowski\AsciiTable\Column;
 use MarcinOrlowski\AsciiTable\ColumnsContainer;
-use MarcinOrlowski\AsciiTable\Exceptions\ColumnKeyNotFound;
+use MarcinOrlowski\AsciiTable\Exceptions\ColumnKeyNotFoundException;
 use MarcinOrlowski\AsciiTable\Output\WriterContract;
 use MarcinOrlowski\AsciiTable\Row;
 use MarcinOrlowski\AsciiTable\Utils\StringUtils;
@@ -29,7 +29,7 @@ class DefaultRenderer implements RendererContract
     /**
      * @inheritDoc
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     public function render(AsciiTable $table, WriterContract $writer): void
     {
@@ -72,7 +72,7 @@ class DefaultRenderer implements RendererContract
      *
      * @return string
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     protected function renderRow(ColumnsContainer $columns, Row $row): string
     {
@@ -119,7 +119,7 @@ class DefaultRenderer implements RendererContract
      *
      * @return string
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     protected function renderHeader(ColumnsContainer $columns): string
     {
@@ -226,7 +226,7 @@ class DefaultRenderer implements RendererContract
      * @param string           $value     Value to pad.
      * @param Align|null       $align     Requested text alignment. If null, column's alignment will be used.
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     protected function pad(ColumnsContainer $columns,
                            string|int       $columnKey,
@@ -262,7 +262,7 @@ class DefaultRenderer implements RendererContract
      * @param ColumnsContainer $columns   Table column definition container.
      * @param string|int       $columnKey Column key we are going to populate.
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     protected function getColumnWidth(ColumnsContainer $columns, string|int $columnKey): int
     {
@@ -275,7 +275,7 @@ class DefaultRenderer implements RendererContract
      * @param ColumnsContainer $columns   Table column definition container.
      * @param string|int       $columnKey Column key we are going to populate.
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     protected function getColumnAlign(ColumnsContainer $columns, string|int $columnKey): Align
     {
@@ -288,7 +288,7 @@ class DefaultRenderer implements RendererContract
      * @param ColumnsContainer $columns   Table column definition container.
      * @param string|int       $columnKey Column key we are going to populate.
      *
-     * @throws ColumnKeyNotFound
+     * @throws ColumnKeyNotFoundException
      */
     protected function getColumnTitleAlign(ColumnsContainer $columns, string|int $columnKey): Align
     {