+++ CHANGES.md
@@ -8,13 +8,14 @@ Fast and flexible PHP library rendering plain ASCII tables.
 
 * @dev
   * Corrected rendering tables with first/last column hidden.
+  * Attempt to render table with all colums hidden throws `NoVisibleColumnsException`.
 
 * 1.6.0 (2022-10-19)
   * Rows can now contain incomplete set of column cells.
 
 
 * 1.5.0 (2022-10-18)
-  * Columns can now be hidden from the rendered output. 
+  * Columns can now be hidden from the rendered output.
 
 
 * 1.4.0 (2022-10-18)