+++ CHANGES.md
@@ -9,6 +9,8 @@ Fast and flexible PHP library rendering plain ASCII tables.
 * @dev
   * Corrected rendering tables with first/last column hidden.
   * Attempt to render table with all colums hidden throws `NoVisibleColumnsException`.
+  * All custom exception class names now ends with `Exception`.
+
 
 * 1.6.0 (2022-10-19)
   * Rows can now contain incomplete set of column cells.